Introduce abstract class FamiliarEntity to avoid code duplication
