+++ src/main/java/com/github/klikli_dev/occultism/common/entity/BatFamiliarEntity.java
@@ -22,14 +22,17 @@
 
 package com.github.klikli_dev.occultism.common.entity;
 
-import net.minecraft.entity.CreatureEntity;
+import com.google.common.collect.ImmutableList;
+
 import net.minecraft.entity.EntityType;
 import net.minecraft.entity.MobEntity;
 import net.minecraft.entity.ai.attributes.AttributeModifierMap;
 import net.minecraft.entity.ai.attributes.Attributes;
+import net.minecraft.potion.EffectInstance;
+import net.minecraft.potion.Effects;
 import net.minecraft.world.World;
 
-public class BatFamiliarEntity extends CreatureEntity {
+public class BatFamiliarEntity extends FamiliarEntity {
 
     public BatFamiliarEntity(EntityType<? extends BatFamiliarEntity> type, World worldIn) {
         super(type, worldIn);
@@ -38,5 +41,10 @@ public class BatFamiliarEntity extends CreatureEntity {
     public static AttributeModifierMap.MutableAttribute registerAttributes() {
         return MobEntity.func_233666_p_().createMutableAttribute(Attributes.MAX_HEALTH, 6)
                 .createMutableAttribute(Attributes.MOVEMENT_SPEED, 0.3);
+    }
+
+    @Override
+    public Iterable<EffectInstance> getFamiliarEffects() {
+        return ImmutableList.of(new EffectInstance(Effects.NIGHT_VISION, 300, 1, false, false));
     }
 }
